Reduce performance hit of recursive mutex
The recursive mutex code seemed to negatively effect some time critical code, so the code was reworked slightly
for the regular mutex methods, but additional explciit mutex_funx_r() and mutex_func_nr() vairants were added for
recursive mutexes and non recursive mutexes respectively.

It might have been nice to swtich to different mutex types for recursive/non recursive (which i had considered
when adding recursive mutexes), but that is a bit too fraught with backwards compatibility issues.

I imagine not _that_ many people care about the performance of this so critically that they much care (which is why the _r and _nr names are a bit oblique - the naming was a bit confusing as they will silently fail in release mode for the wrong type of mutex)