Resolve __builtin_memcpy for non-GCC compilers
IAR generates warning Pe223 'function "__builtin_memcpy" declared implicitly'
and then error Li005 'no definition for "__builtin_memcpy"' at link time.

It appears from a disassembly of the binary built by gcc 9.2.1 that this
only resolving to a call to memcpy anyway, so it's not clear what the
benefit in using __builtin_memcpy was - but I've left GCC calling it to be
on the safe side.