Ensure binary info pointers go in flash
These weren't declared const, which means the IAR linker ends up splitting
the sections into an initialiser section (with "_init" suffix) and a working
section (without). But the latter can't both be allocated in flash and
given contents. By inserting the missing "const" qualifier, we only have one
section (a read-only one) to deal with.