Provide GCC-compatible definition  __builtin_ctz for IAR builds
IAR generates warning Pe223 'function "__builtin_ctz" declared implicitly',
though it does link successfully without this change. This is because the
IAR runtime defines the function, despite it not being declared in any of
its header files.

This inline function should boil down to a call to __ctzsi2, which in turn
calls the implementation in the boot ROM, rather than the one in the IAR
runtime library.