Add condition variables to pico_sync (fix #1093)
Implement condition variables as a companion to mutexes. Condition variables
can be signaled and broadcast and implementation should work with any number
of cores.

To prevent deadlocks, at most a single spin lock is held at a given time. As a
result there can be a race condition if several cores call cond_signal at the
same time (without holding the mutex).