Restore "API Documentation" in the main menu
Between Doxygen 1.8.17 and Doxygen 1.9.8, the top-level "modules" tab
type in the Doxygen layout has changed such that our "API Documentation"
section is no longer generated and is instead considered of the type
"topics". Update our layout XML to reflect this change and restore this
critical section of the documentation.

Fixes #1563.