Sets PICO_DEOPTIMIZED_DEBUG to be enabled by default
Optimizing a debug build with PICO_DEOPTIMIZED DEBUG disabled (the current default) results in breakpoints not being possible in certain functions.  This change makes the default behaviour of PICO_DEOPTIMIZED_DEBUG to be enabled, so that debug builds will fully work as expected when no explicit value for the option is passed.

Since the cache stores the default value of PICO_DEOPTIMIZED DEBUG, you will need to delete the CMakeCache.txt previously generated in your project directory for this change to default behaviour to take effect in existing projects.

This commit is related to pull request raspberrypi#1620 in the primary repository.

See Issue raspberrpi#1618 in the primary repository for additional context.