Bug 1668: Provide method to prevent init of SYS_CLK, when left on during 'lightsleep'
During 'lightsleep' user may want to keep some clocks running (for example PIOs),
but at present microPython calls 'clock-init()' causing the SYS_CLK to be reinitialised.

This patch allows microPython to provide the previously used 'sleep_en0 & sleep_en1'
signifying which clocks stayed on. As most are dependant of SYS_CLK it can be
inferred that this does NOT need to (/should NOT) be re-initilised.