Change BT flash storage default for rp2350 A2
The workaround for errata RP2350-E10 overwrites the last block in flash.
This will overwrite the BT flash storage causing a paired BT connection
to fail. Move the default flash storage location to 3 sectors from the
end of flash for RP2350 where A2 support is required.

This will require existing BT pairings to a Pico device to be removed
and readded.

Fixes #2322