Flash functions in PICO_NO_FLASH binaries
We assume that PICO_NO_FLASH binaries means the device does not have
a flash chip. But it's possible that you want a ram binary AND want
to use the unique flash id functionality.

Add PICO_INCLUDE_FLASH_ID_FUNCTIONS to force the inclusion of these
flash functions in ram binaries.

Fixes #2796