Allow 'host' platform to simulate either RP2040 or RP2350
The 'host' platform only really partially simulates an RP2040
environment, with no way to specify a particular board either, so
attempting to build a host version of a existing code that uses an
RP2350 or relies on values defined in a board file is either difficult
or impossible.

Instead of setting PICO_PLATFORM to just 'host', it can now be set to
'host-rp2040' or 'host-rp2350' and PICO_BOARD can be set to any board
configuration that is compatible with the given chip type.
Platform definitions for the given chip will be fetched, meaning
defines such as 'NUM_IRQS' will be set correctly.

Fixes #2778